* Little update ¯\_(ツ)_/¯

* Revert "Little update ¯\_(ツ)_/¯"

This reverts commit c1e4608a16bb990b53dbae0481ac274fab94af74.

* Big Boy

* Reobfuscated Mixins

* Reobfuscated Mixins

* Imports.

* Reobfuscated Mixins

* Updated imports, names and probably fixed some things. Removed mg.java.

* Fixed Imports

* Do all the imports & cleanup.

+++ src/main/java/io/github/ryuu/adventurecraft/entities/EntitySkeletonSword.java
@@ -7,6 +7,7 @@ import net.minecraft.item.ItemType;
 import net.minecraft.level.Level;
 
 public class EntitySkeletonSword extends Skeleton {
+
     public EntitySkeletonSword(Level world) {
         super(world);
         this.attackDamage = 1;
@@ -15,7 +16,7 @@ public class EntitySkeletonSword extends Skeleton {
 
     @Override
     protected void method_637(Entity entity, float f) {
-        if (f < 2.5D && entity.boundingBox.maxY > this.boundingBox.minY && entity.boundingBox.minY < this.boundingBox.maxY) {
+        if ((double) f < 2.5 && entity.boundingBox.maxY > this.boundingBox.minY && entity.boundingBox.minY < this.boundingBox.maxY) {
             this.attackTime = 20;
             entity.damage(this, this.attackDamage);
         }
